Fixed build errors with gcc 4.7.0

+++ libs/MAUtil/mauuid.h
@@ -26,7 +26,7 @@ Software Foundation, 59 Temple Place - Suite 330, Boston, MA
 * Defines a MAUUID that follows the Bluetooth standard of UUIDs.
 * The lower 96 bits are constant.
 */
-#define DEFINE_BTMAUUID(name, uuid16) static const MAUUID name = {{ uuid16, 0x00001000, 0x80000080, 0x5F9B34FB }}
+#define DEFINE_BTMAUUID(name, uuid16) static const MAUUID name = {{ uuid16, 0x00001000, (int)0x80000080, 0x5F9B34FB }}
 
 DEFINE_BTMAUUID(Bluetooth_Base_MAUUID, 0);
